+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Fixed
+- Added missing benchmark arguments to the `Benchmarker.benchmark` method.
 - Fixed another issue related to the `download_only` mode, causing model evaluations to
   fail, as it could not find the model locally. This has been fixed now.